Port code into vcf_utils file
